Refactored to remove code in loop and moved to callbacks instead.
I was still having trouble with the board losing connection and requiring a particle device doctor reset each time. My new theory is that this was due to using delays in the main loop resulting in the board losing communication with the blynk server. The docs seemed to imply this could be a possibility. This change therefore removes all delays from the main loop (seems inefficient but that's what the examples suggest), and added items to the callbacks instead.