Measures system temperature via the ADC on A0.
This uses a BlynkTimer to trigger the measurement and calculation. This seemed to be the recommended way of doing this so that the main loop can spin at a much higher frequency. When the timer gets triggered it'll inevitably delay the main loop, but apparently (at least I haven't seen a problem yet), it seems to be happy.