Add class to control the output of the system.
The controller uses a basic hysterisis to compare the setpoint and measured temp to decide whether to switch on the heating or not. The controller is called at startup, when a setpoint change is made and when the measured temp is read.
Implements a hold-off timer to ensure the output doesn't get switched too quickly.

In future, if a more sophisticated control system is required, this could be implemented within this Controller structure.